+++ src/main/java/org/wso2/carbon/connector/operations/EmailExpungeFolder.java
@@ -27,10 +27,9 @@ import org.wso2.carbon.connector.exception.ContentBuilderException;
 import org.wso2.carbon.connector.exception.EmailConnectionException;
 import org.wso2.carbon.connector.exception.EmailConnectionPoolException;
 import org.wso2.carbon.connector.exception.InvalidConfigurationException;
-import org.wso2.carbon.connector.utils.ConfigurationUtils;
 import org.wso2.carbon.connector.utils.EmailConstants;
+import org.wso2.carbon.connector.utils.EmailUtils;
 import org.wso2.carbon.connector.utils.Error;
-import org.wso2.carbon.connector.utils.ResponseHandler;
 
 import javax.mail.Folder;
 
@@ -44,31 +43,31 @@ public class EmailExpungeFolder extends AbstractConnector {
     @Override
     public void connect(MessageContext messageContext) {
 
-        String errorString = "Error occurred while expunging folder: %s. %s";
+        String errorString = "Error occurred while expunging folder: %s.";
         String folder = (String) getParameter(messageContext, EmailConstants.FOLDER);
         EmailConnectionPool pool = null;
         MailBoxConnection connection = null;
         try {
-            if (StringUtils.isEmpty(folder)){
+            if (StringUtils.isEmpty(folder)) {
                 folder = EmailConstants.DEFAULT_FOLDER;
             }
-            String connectionName = ConfigurationUtils.getConnectionName(messageContext);
+            String connectionName = EmailUtils.getConnectionName(messageContext);
             pool = EmailConnectionManager.getEmailConnectionManager().getConnectionPool(connectionName);
             connection = (MailBoxConnection) pool.borrowObject();
             expungeFolder(connection, folder);
             if (log.isDebugEnabled()) {
                 log.debug(format("Expunged folder: %s...", folder));
             }
-            ResponseHandler.generateOutput(messageContext, true);
+            EmailUtils.generateOutput(messageContext, true);
         } catch (EmailConnectionException | EmailConnectionPoolException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.CONNECTIVITY);
-            handleException(format(errorString, folder, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.CONNECTIVITY);
+            handleException(format(errorString, folder), e, messageContext);
         } catch (InvalidConfigurationException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.INVALID_CONFIGURATION);
-            handleException(format(errorString, folder, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.INVALID_CONFIGURATION);
+            handleException(format(errorString, folder), e, messageContext);
         } catch (ContentBuilderException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.RESPONSE_GENERATION);
-            handleException(format(errorString, folder, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.RESPONSE_GENERATION);
+            handleException(format(errorString, folder), e, messageContext);
         } finally {
             if (pool != null) {
                 pool.returnObject(connection);
