+++ src/main/java/org/wso2/carbon/connector/operations/EmailMarkAsDeleted.java
@@ -27,11 +27,9 @@ import org.wso2.carbon.connector.exception.EmailConnectionException;
 import org.wso2.carbon.connector.exception.EmailConnectionPoolException;
 import org.wso2.carbon.connector.exception.EmailNotFoundException;
 import org.wso2.carbon.connector.exception.InvalidConfigurationException;
-import org.wso2.carbon.connector.utils.ConfigurationUtils;
 import org.wso2.carbon.connector.utils.EmailConstants;
-import org.wso2.carbon.connector.utils.EmailOperationUtils;
+import org.wso2.carbon.connector.utils.EmailUtils;
 import org.wso2.carbon.connector.utils.Error;
-import org.wso2.carbon.connector.utils.ResponseHandler;
 
 import javax.mail.Flags;
 
@@ -47,28 +45,28 @@ public class EmailMarkAsDeleted extends AbstractConnector {
 
         String folder = (String) getParameter(messageContext, EmailConstants.FOLDER);
         String emailID = (String) getParameter(messageContext, EmailConstants.EMAIL_ID);
-        String errorString = "Error occurred while marking email with ID: %s as deleted. %s";
+        String errorString = "Error occurred while marking email with ID: %s as deleted.";
         EmailConnectionPool pool = null;
         MailBoxConnection connection = null;
         try {
-            String connectionName = ConfigurationUtils.getConnectionName(messageContext);
+            String connectionName = EmailUtils.getConnectionName(messageContext);
             pool = EmailConnectionManager.getEmailConnectionManager().getConnectionPool(connectionName);
             connection = (MailBoxConnection) pool.borrowObject();
-            boolean status = EmailOperationUtils.changeEmailState(connection, folder, emailID,Flags.Flag.DELETED,
+            boolean status = EmailUtils.changeEmailState(connection, folder, emailID, Flags.Flag.DELETED,
                     false);
-            ResponseHandler.generateOutput(messageContext, status);
+            EmailUtils.generateOutput(messageContext, status);
         } catch (EmailConnectionException | EmailConnectionPoolException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.CONNECTIVITY);
-            handleException(format(errorString, emailID, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.CONNECTIVITY);
+            handleException(format(errorString, emailID), e, messageContext);
         } catch (EmailNotFoundException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.EMAIL_NOT_FOUND);
-            handleException(format(errorString, emailID, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.EMAIL_NOT_FOUND);
+            handleException(format(errorString, emailID), e, messageContext);
         } catch (InvalidConfigurationException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.INVALID_CONFIGURATION);
-            handleException(format(errorString, folder, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.INVALID_CONFIGURATION);
+            handleException(format(errorString, folder), e, messageContext);
         } catch (ContentBuilderException e) {
-            ResponseHandler.setErrorsInMessage(messageContext, Error.RESPONSE_GENERATION);
-            handleException(format(errorString, folder, e.getMessage()), e, messageContext);
+            EmailUtils.setErrorsInMessage(messageContext, Error.RESPONSE_GENERATION);
+            handleException(format(errorString, folder), e, messageContext);
         } finally {
             if (pool != null) {
                 pool.returnObject(connection);
