--- src/main/java/org/wso2/carbon/connector/utils/ContentTypes.java
@@ -1,35 +1,0 @@
-/*
- *  Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.wso2.carbon.connector.utils;
-
-/**
- * Contains the content types
- */
-public final class ContentTypes {
-
-    // Content Types
-    public static final String APPLICATION_XML = "application/xml";
-    public static final String APPLICATION_JSON = "application/json";
-    public static final String TEXT_XML = "text/xml";
-    public static final String TEXT_CSV = "text/csv";
-    public static final String TEXT_PLAIN = "text/plain";
-    private ContentTypes() {
-
-    }
-
-}
